+++ src/tunnel-services.ts
@@ -1,5 +1,6 @@
 import childProcess, {SpawnOptions} from 'child_process'
 import {createServer, Socket} from 'net'
+import {Client as PgClient, ClientConfig as PgClientConfig} from 'pg'
 import {Client as SshClient} from 'ssh2'
 
 /**
@@ -12,6 +13,7 @@ export default {
     spawn: (command: string, options: SpawnOptions) => childProcess.spawn(command, options),
   },
   nodeProcess: process,
+  pgClientFactory: {create: (config: PgClientConfig) => new PgClient(config)},
   sshClientFactory: {create: () => new SshClient()},
   tcpServerFactory: {
     create: (connectionListener: (socket: Socket) => void) => createServer(connectionListener),