+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 This project adheres to [Semantic Versioning](http://semver.org/). All notable changes will be documented in this file.
 
 ## [Unreleased](https://github.com/OldSneerJaw/borealis-pg-cli/compare/v0.1.0...HEAD)
-- `borealis-pg:run` command to execute a noninteractive shell command for an add-on database
+- `borealis-pg:run` command to execute a noninteractive SQL or shell command for an add-on database
 
 ## [0.1.0](https://github.com/OldSneerJaw/borealis-pg-cli/compare/477321d...v0.1.0) - 2021-09-24
 First public pre-release version